requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.9.3
blis==0.7.6
catalogue==1.0.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cymem==2.0.6
errant==2.3.0
ffmpy3==0.2.4
filelock==3.6.0
Flask==2.0.3
frozenlist==1.3.0
fsspec==2022.2.0
future==0.18.2
fuzzywuzzy==0.18.0
gramformer @ git+https://github.com/PrithivirajDamodaran/Gramformer.git@41e396937a122e10294e04069ad89430f39fa778
gTTS==2.2.4
huggingface-hub==0.4.0
HyperPyYAML==1.0.0
idna==2.10
importlib-resources==5.4.0
itsdangerous==2.1.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter-core==4.9.2
line-bot-sdk==2.1.0
MarkupSafe==2.1.1
multidict==6.0.2
murmurhash==1.0.6
nbformat==5.2.0
nnsplit==0.5.8.post0
numpy==1.22.3
onnxruntime==1.7.0
packaging==21.3
pigar==0.10.0
plac==1.1.3
preshed==3.0.6
protobuf==3.19.4
pyparsing==3.0.7
pyrsistent==0.18.1
python-dotenv==0.19.2
python-Levenshtein==0.12.2
PyYAML==6.0
regex==2022.3.15
requests==2.25.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sacremoses==0.0.49
scipy==1.8.0
selenium==3.141.0
sentencepiece==0.1.96
six==1.16.0
soupsieve==2.0.1
spacy==2.3.7
speechbrain==0.5.11
srsly==1.0.5
thinc==7.4.5
tokenizers==0.11.6
torch==1.11.0
torchaudio==0.11.0
tqdm==4.63.0
traitlets==5.1.1
transformers==4.17.0
typing_extensions==4.1.1
urllib3==1.26.2
wasabi==0.9.0
Werkzeug==2.0.3
yarl==1.7.2
zipp==3.7.0
